# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DebtManager API",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Customer"
              }
            }
          }
        }
      }
    },
    "/api/Email/send-reset-token": {
      "post": {
        "tags": [
          "Email"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmailRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Gemini/generate": {
      "post": {
        "tags": [
          "Gemini"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GeminiRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Gemini/analyze-receipt": {
      "post": {
        "tags": [
          "Gemini"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GeminiRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Gemini/bulk-add": {
      "post": {
        "tags": [
          "Gemini"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GeminiRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AnalyticsCustomerStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "Category": {
      "required": [
        "name",
        "organizationId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "parentCategoryId": {
          "format": "int32",
          "type": "integer"
        },
        "parentCategory": {
          "$ref": "#/definitions/Category"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "deletedAt": {
          "format": "date-time",
          "type": "string"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        },
        "syncStatus": {
          "$ref": "#/definitions/SyncStatus"
        },
        "lastModified": {
          "format": "date-time",
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "CurrencyType": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "Customer": {
      "required": [
        "name",
        "organizationId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phoneNumber": {
          "maxLength": 20,
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "deletedAt": {
          "format": "date-time",
          "type": "string"
        },
        "debtLimit": {
          "format": "double",
          "type": "number"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        },
        "dateAdded": {
          "format": "date-time",
          "type": "string"
        },
        "inactivityPeriodDays": {
          "format": "int32",
          "type": "integer"
        },
        "syncStatus": {
          "$ref": "#/definitions/SyncStatus"
        },
        "lastModified": {
          "format": "date-time",
          "type": "string"
        },
        "transactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Transaction"
          }
        },
        "sales": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Sale"
          }
        },
        "analyticsStatus": {
          "$ref": "#/definitions/AnalyticsCustomerStatus"
        },
        "status": {
          "$ref": "#/definitions/AnalyticsCustomerStatus"
        }
      }
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "DayOfWeek": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
    "EmailRequestDto": {
      "required": [
        "body",
        "subject",
        "to"
      ],
      "type": "object",
      "properties": {
        "to": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "subject": {
          "minLength": 1,
          "type": "string"
        },
        "body": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "GeminiRequestDto": {
      "type": "object",
      "properties": {
        "prompt": {
          "type": "string"
        }
      }
    },
    "Organization": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 256,
          "minLength": 1,
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "customers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Customer"
          }
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "sales": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Sale"
          }
        },
        "transactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Transaction"
          }
        },
        "settings": {
          "$ref": "#/definitions/OrganizationSettings"
        }
      }
    },
    "OrganizationSettings": {
      "required": [
        "organizationId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "shopName": {
          "type": "string"
        },
        "logoPath": {
          "type": "string"
        },
        "overdueDays": {
          "format": "int32",
          "type": "integer"
        },
        "theme": {
          "type": "string"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        }
      }
    },
    "Product": {
      "required": [
        "name",
        "organizationId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 200,
          "minLength": 0,
          "type": "string"
        },
        "sku": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "costPrice": {
          "format": "double",
          "type": "number"
        },
        "sellingPrice": {
          "format": "double",
          "type": "number"
        },
        "stockQuantity": {
          "format": "double",
          "type": "number"
        },
        "lowStockThreshold": {
          "format": "double",
          "type": "number"
        },
        "isActive": {
          "type": "boolean"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "deletedAt": {
          "format": "date-time",
          "type": "string"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "supplier": {
          "type": "string"
        },
        "barcode": {
          "type": "string"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        },
        "syncStatus": {
          "$ref": "#/definitions/SyncStatus"
        },
        "lastModified": {
          "format": "date-time",
          "type": "string"
        },
        "saleItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SaleItem"
          }
        }
      }
    },
    "Sale": {
      "required": [
        "organizationId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "saleDate": {
          "format": "date-time",
          "type": "string"
        },
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "isPaid": {
          "type": "boolean"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "deletedAt": {
          "format": "date-time",
          "type": "string"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        },
        "syncStatus": {
          "$ref": "#/definitions/SyncStatus"
        },
        "lastModified": {
          "format": "date-time",
          "type": "string"
        },
        "saleItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SaleItem"
          }
        }
      }
    },
    "SaleItem": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "saleId": {
          "format": "int32",
          "type": "integer"
        },
        "sale": {
          "$ref": "#/definitions/Sale"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/Product"
        },
        "quantity": {
          "format": "double",
          "type": "number"
        },
        "unitPrice": {
          "format": "double",
          "type": "number"
        },
        "totalPrice": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      }
    },
    "SyncStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "Transaction": {
      "required": [
        "amount",
        "currency",
        "customerId",
        "organizationId",
        "type"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "customerId": {
          "format": "int32",
          "type": "integer"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "type": {
          "$ref": "#/definitions/TransactionType"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "deletedAt": {
          "format": "date-time",
          "type": "string"
        },
        "currency": {
          "$ref": "#/definitions/CurrencyType"
        },
        "receiptId": {
          "type": "string"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/Product"
        },
        "quantity": {
          "format": "double",
          "type": "number"
        },
        "syncStatus": {
          "$ref": "#/definitions/SyncStatus"
        },
        "lastModified": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "TransactionType": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "User": {
      "required": [
        "passwordHash",
        "role",
        "username"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "passwordHash": {
          "minLength": 1,
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/UserRole"
        },
        "email": {
          "maxLength": 255,
          "type": "string"
        },
        "passwordResetToken": {
          "type": "string"
        },
        "resetTokenExpiry": {
          "format": "date-time",
          "type": "string"
        },
        "organizationId": {
          "format": "int32",
          "type": "integer"
        },
        "organization": {
          "$ref": "#/definitions/Organization"
        }
      }
    },
    "UserLoginDto": {
      "required": [
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UserRegisterDto": {
      "required": [
        "email",
        "organizationName",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "maxLength": 100,
          "minLength": 3,
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "minLength": 6,
          "type": "string"
        },
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "organizationName": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UserRole": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "$ref": "#/definitions/DateOnly"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. \r\n                      Enter 'Bearer' [space] and then your token in the text input below.\r\n                      Example: 'Bearer 12345abcdef'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}